+++ models/staging/camara_deputados/stg_camara_temas.sql
@@ -11,7 +11,7 @@ renamed AS (
     SELECT
         SPLIT_PART(url_temas, '/', 7)      AS proposicao_id_nk,
         {{ clean_string("tema","upper") }} AS tema,
-        codtema                            AS codigo_tema,
+        codtema::BIGINT                    AS codigo_tema,
         relevancia::INT                    AS relevancia
     FROM source
 )
