+++ models/staging/camara_deputados/stg_camara_votos_deputados.sql
@@ -13,7 +13,7 @@ renamed AS (
         deputado__nome                                AS nome,
         deputado__siglauf                             AS uf,
         {{ clean_string("tipovoto","upper") }}        AS voto,
-        deputado__idlegislatura                       AS legislatura_id_fk,
+        deputado__idlegislatura::BIGINT               AS legislatura_id_fk,
         SPLIT_PART(url_votos, '/', 7)                 AS votacao_id_fk,
         SPLIT_PART(deputado__uripartido, '/', 7)::INT AS partido_id_fk
     FROM source
